パスワードリセット機能を改善し、リセットメール送信のロジックを修正。ResetPasswordConfirmコンポーネントを削除し、UIを整理。関連するメッセージを更新。
