+++ pro/src/pages/PasswordReset.jsx
@@ -1,23 +1,32 @@
 import React, { useState } from "react";
-import { auth } from "../firebaseConfig"; // Firebaseの初期化ファイルをインポート
+import { auth } from "../firebaseConfig";
 import { sendPasswordResetEmail } from "firebase/auth";
-import { TextField, Button, Container, Typography, Box, Alert } from "@mui/material";
+import { Button, Container, Typography, Box, Alert } from "@mui/material";
 import Layout from "../components/Layout";
 
-const PasswordReset = () => {
-  const [email, setEmail] = useState("");
+const PasswordReset = () => { // コンポーネント名を修正
   const [message, setMessage] = useState("");
   const [error, setError] = useState("");
 
-  const handlePasswordReset = async () => {
+  const handleSendResetEmail = async () => {
     setMessage("");
     setError("");
+
+    const user = auth.currentUser;
+
+    if (!user || !user.email) {
+      setError("ログインしているユーザー情報が取得できませんでした。再ログインしてください。");
+      return;
+    }
+
     try {
-      await sendPasswordResetEmail(auth, email);
-      setMessage("パスワードリセットメールを送信しました。メールボックスを確認してください。");
+      await sendPasswordResetEmail(auth, user.email);
+      setMessage(
+        `パスワードリセットメールを登録済みのメールアドレス (${user.email}) に送信しました。メールを確認してください。`
+      );
     } catch (err) {
-      setError("パスワードリセットに失敗しました。正しいメールアドレスを入力してください。");
-      console.error("パスワードリセットエラー:", err);
+      setError("メール送信に失敗しました。サポートにお問い合わせください。");
+      console.error("パスワードリセットメール送信エラー:", err);
     }
   };
 
@@ -26,25 +35,19 @@ const PasswordReset = () => {
       <Container maxWidth="sm" sx={{ mt: 4 }}>
         <Box sx={{ p: 3, boxShadow: 2, borderRadius: 2, textAlign: "center" }}>
           <Typography variant="h5" gutterBottom>
-            パスワードリセット
+            パスワードリセットメール送信
           </Typography>
-          <TextField
-            label="登録済みのメールアドレス"
-            type="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            variant="outlined"
-            fullWidth
-            margin="normal"
-          />
+          <Typography variant="body1" gutterBottom>
+            登録済みのメールアドレスにリセットメールを送信します。
+          </Typography>
           <Button
             variant="contained"
             color="primary"
-            onClick={handlePasswordReset}
+            onClick={handleSendResetEmail}
             fullWidth
             sx={{ mt: 2 }}
           >
-            パスワードリセットメールを送信
+            リセットメールを送信
           </Button>
           {message && (
             <Alert severity="success" sx={{ mt: 2 }}>
